refactor: move global hotkey logic to emain-window
