+++ emain/emain.ts
@@ -28,7 +28,7 @@ import {
     setWasInFg,
 } from "./emain-activity";
 import { ensureHotSpareTab, getWaveTabViewByWebContentsId, setMaxTabCacheSize } from "./emain-tabview";
-import { handleCtrlShiftState, waveKeyToElectronKey } from "./emain-util";
+import { handleCtrlShiftState } from "./emain-util";
 import { getIsWaveSrvDead, getWaveSrvProc, getWaveSrvReady, getWaveVersion, runWaveSrv } from "./emain-wavesrv";
 import {
     createBrowserWindow,
@@ -38,6 +38,7 @@ import {
     getWaveWindowById,
     getWaveWindowByWebContentsId,
     getWaveWindowByWorkspaceId,
+    registerGlobalHotkey,
     relaunchBrowserWindows,
     WaveBrowserWindow,
 } from "./emain-window";
@@ -612,24 +613,7 @@ async function appMain() {
     });
     const rawGlobalHotKey = launchSettings?.["key:globalhotkey"];
     if (rawGlobalHotKey) {
-        try {
-            const electronHotKey = waveKeyToElectronKey(rawGlobalHotKey);
-            console.log("registering globalhotkey of ", electronHotKey);
-            electron.globalShortcut.register(electronHotKey, () => {
-                const selectedWindow = focusedWaveWindow;
-                const firstWaveWindow = getAllWaveWindows()[0];
-                console.log("the selected window is ", selectedWindow ?? "nonexistent");
-                if (focusedWaveWindow) {
-                    selectedWindow.focus();
-                } else if (firstWaveWindow) {
-                    firstWaveWindow.focus();
-                } else {
-                    fireAndForget(createNewWaveWindow);
-                }
-            });
-        } catch (e) {
-            console.log(e);
-        }
+        registerGlobalHotkey(rawGlobalHotKey);
     }
 }
